refactor(server): extract bot AI from OrbitalLobbyRoom, remove dead botBounceArena code
OrbitalLobbyRoom.ts mixed lobby flow, match flow, the bot brain, and
geometry helpers. Move the bot brain (personality, launch/fire timers,
per-tick drift, targeting, firing) into colyseus/botAI.ts as a
BotController that triggers match-level effects through BotCombatHooks.
The freeze-on-hit bookkeeping that was duplicated between the bot path
and handleHitReportMessage is now one applyFreezeFromShot helper.

botBounceArena had an early 'return;' followed by ~50 lines of
unreachable code — the live behavior is bounceActorInArena, which the
bot brain now calls directly. botIsInEnemyBreachRoom duplicated
isActorInEnemyBreachRoom and is gone too; breachRoomCenter moves to
onlineActorSimulation.ts so both modules share it.

OrbitalLobbyRoom shrinks from ~1290 to ~1036 lines. New pure functions
(botPersonality, botIdHash) get vitest coverage; the bot-targeting
tests now import from botAI.
