+++ server/src/colyseus/OrbitalLobbyRoom.ts
@@ -42,10 +42,12 @@ import {
   PLAYER_RADIUS,
 } from "../../../shared/constants";
 import type { PlayerPhase } from "../../../shared/schema";
-import { generateSpawnPositions, resolveActorCollisions, type CollisionBody } from "../../../shared/player-logic";
+import { generateSpawnPositions, resolveActorCollisions, type CollisionBody, type HitZone } from "../../../shared/player-logic";
 import { applyHitToOnlineActor, isHitZone, normalizeAuthoritativePhase } from "./actorDamage";
+import { BotController, botIdHash, botPersonality, type BotCombatHooks } from "./botAI";
 import {
   bounceActorInArena,
+  breachRoomCenter,
   integrateZeroGActor,
   isActorInEnemyBreachRoom,
   isHitReportDistancePlausible,
@@ -83,10 +85,15 @@ export class OrbitalLobbyRoom extends Room<{ state: OrbitalLobbyState }> {
   private matchTick: ReturnType<typeof setInterval> | null = null;
   private botCounters: Record<LobbyTeam, number> = { 0: 0, 1: 0 };
   private botSpawnYaw: Record<LobbyTeam, number> = { 0: 0, 1: 0 };
-  private botAI = new Map<string, { launchTimer: number }>();
+  private bots = new BotController<ActorState>();
   private botGoalAxis: "x" | "z" = "x";
   private botGoalSigns: { team0: 1 | -1; team1: 1 | -1 } = { team0: 1, team1: -1 };
-  private botFireTimers = new Map<string, number>();
+  private readonly botHooks: BotCombatHooks<ActorState> = {
+    applyFreeze: (shooter, target) => this.applyFreezeFromShot(shooter, target, "body"),
+    awardBreachPoint: (bot) => this.awardOnlineRoundPoint(bot.team, bot.id, bot.name, "breach"),
+    broadcastShot: (event) => this.broadcast("shot_event", event),
+    isRoundResolved: () => this.roundResolved,
+  };
   private countdownPreparedRound = false;
   private roundResolved = false;
   private lastPlayerUpdate = new Map<string, number>();
@@ -378,14 +385,24 @@ export class OrbitalLobbyRoom extends Room<{ state: OrbitalLobbyState }> {
     if (!isHitZone(message.zone)) return;
     if (!isHitReportDistancePlausible(shooter, target)) return;
 
+    this.applyFreezeFromShot(shooter, target, message.zone);
+  }
+
+  /**
+   * Freeze `target` from a shot by `shooter` (bots freeze instantly,
+   * humans take zone damage), credit the kill, broadcast the freeze
+   * event, and check the full-freeze win. Returns false when the hit
+   * damaged a limb without freezing.
+   */
+  private applyFreezeFromShot(shooter: ActorState, target: ActorState, zone: HitZone): boolean {
     if (target.isBot) {
       target.frozen = true;
       target.phase = "FROZEN";
       target.deaths += 1;
     } else {
-      const frozen = applyHitToOnlineActor(target, message.zone);
+      const frozen = applyHitToOnlineActor(target, zone);
       if (!frozen) {
-        return;
+        return false;
       }
     }
 
@@ -402,6 +419,7 @@ export class OrbitalLobbyRoom extends Room<{ state: OrbitalLobbyState }> {
     this.broadcast("freeze_event", freezeEvent);
 
     this.checkFullFreezeWin();
+    return true;
   }
 
   private handleBreachReportMessage(client: RoomClient): void {
@@ -701,10 +719,7 @@ export class OrbitalLobbyRoom extends Room<{ state: OrbitalLobbyState }> {
       }
 
       if (member.isBot) {
-        const idHash = botIdHash(member.id);
-        const p = botPersonality(idHash);
-        this.botAI.set(member.id, { launchTimer: 1.5 + (idHash % 30) * 0.1 });
-        this.botFireTimers.set(member.id, p.fireDelay * (0.5 + (idHash % 10) * 0.05));
+        this.bots.registerBot(member.id);
       }
 
       this.state.actors.set(member.id, actor);
@@ -713,72 +728,20 @@ export class OrbitalLobbyRoom extends Room<{ state: OrbitalLobbyState }> {
 
   private clearActors(): void {
     this.state.actors.clear();
-    this.botAI.clear();
-    this.botFireTimers.clear();
+    this.bots.clear();
     this.lastPlayerUpdate.clear();
   }
 
   private removePresence(id: string): void {
     this.state.actors.delete(id);
-    this.botAI.delete(id);
-    this.botFireTimers.delete(id);
+    this.bots.remove(id);
     this.lastPlayerUpdate.delete(id);
   }
 
+
   private tickBots(dt: number): void {
     if (this.state.phase !== "PLAYING") return;
-
-    for (const actor of this.state.actors.values()) {
-      if (!actor.isBot) continue;
-      if (actor.frozen) continue; // stay frozen until round end
-
-      const idHash = botIdHash(actor.id);
-      const p = botPersonality(idHash);
-
-      if (actor.phase === "BREACH") {
-        const ai = this.botAI.get(actor.id);
-        if (!ai) continue;
-        ai.launchTimer -= dt;
-        if (ai.launchTimer <= 0) {
-          const dx = -actor.posX + (Math.random() - 0.5) * p.angleNoise * 6;
-          const dy = -actor.posY + (Math.random() - 0.5) * p.angleNoise * 3;
-          const dz = -actor.posZ + (Math.random() - 0.5) * p.angleNoise * 6;
-          const len = Math.hypot(dx, dy, dz) || 1;
-          actor.velX = (dx / len) * p.launchSpeed;
-          actor.velY = (dy / len) * p.launchSpeed;
-          actor.velZ = (dz / len) * p.launchSpeed;
-          actor.phase = "FLOATING";
-          const horizLen = Math.hypot(dx, dz);
-          if (horizLen > 0.01) {
-            actor.yaw = Math.atan2(-dx / horizLen, -dz / horizLen);
-          }
-        }
-        continue;
-      }
-
-      if (actor.phase === "FLOATING") {
-        botIntegrateZeroG(actor, dt);
-        botBounceArena(actor, this.botGoalAxis);
-
-        const horizSpeed = Math.hypot(actor.velX, actor.velZ);
-        if (horizSpeed > 0.5) {
-          actor.yaw = Math.atan2(-actor.velX, -actor.velZ);
-        }
-
-        if (!this.roundResolved && this.botIsInEnemyBreachRoom(actor)) {
-          this.awardOnlineRoundPoint(actor.team, actor.id, actor.name, "breach");
-        }
-
-        const fireTimer = this.botFireTimers.get(actor.id) ?? p.fireDelay;
-        const nextFireTimer = fireTimer - dt;
-        if (nextFireTimer <= 0) {
-          this.botTryFire(actor, p);
-          this.botFireTimers.set(actor.id, p.fireDelay * (0.8 + Math.random() * 0.4));
-        } else {
-          this.botFireTimers.set(actor.id, nextFireTimer);
-        }
-      }
-    }
+    this.bots.tick(dt, this.state.actors.values(), this.botGoalAxis, this.botGoalSigns, this.botHooks);
   }
 
   private tickStaleHumanActors(dt: number): void {
@@ -831,88 +794,6 @@ export class OrbitalLobbyRoom extends Room<{ state: OrbitalLobbyState }> {
         actor.velZ = body.vel.z;
       }
     }
-  }
-
-  private botIsInEnemyBreachRoom(bot: ActorState): boolean {
-    const enemyTeam = (bot.team === 0 ? 1 : 0) as 0 | 1;
-    const enemySign = enemyTeam === 0 ? this.botGoalSigns.team0 : this.botGoalSigns.team1;
-    const goalAxis = this.botGoalAxis;
-    const perpAxis: "x" | "z" = goalAxis === "x" ? "z" : "x";
-    const botOnGoal = goalAxis === "x" ? bot.posX : bot.posZ;
-    const botOnPerp = perpAxis === "x" ? bot.posX : bot.posZ;
-    const arenaEdge = enemySign * (ARENA_SIZE / 2);
-    const roomBack = enemySign * (ARENA_SIZE / 2 + BREACH_ROOM_D);
-    const inDepth = enemySign > 0
-      ? botOnGoal > arenaEdge && botOnGoal < roomBack
-      : botOnGoal < arenaEdge && botOnGoal > roomBack;
-    return inDepth && Math.abs(bot.posY) < BREACH_ROOM_H / 2 && Math.abs(botOnPerp) < BREACH_ROOM_W / 2;
-  }
-
-  private findNearestBotEnemy(bot: ActorState): ActorState | null {
-    let nearest: ActorState | null = null;
-    let nearestDistSq = Infinity;
-    for (const actor of this.state.actors.values()) {
-      if (!isOnlineActorTargetableByBot(bot.team, actor, this.botGoalAxis, this.botGoalSigns)) continue;
-      const dx = actor.posX - bot.posX;
-      const dy = actor.posY - bot.posY;
-      const dz = actor.posZ - bot.posZ;
-      const distSq = dx * dx + dy * dy + dz * dz;
-      if (distSq < nearestDistSq) {
-        nearestDistSq = distSq;
-        nearest = actor;
-      }
-    }
-    return nearest;
-  }
-
-  private botTryFire(bot: ActorState, p: BotPersonality): void {
-    if (bot.rightArm || bot.frozen) return;
-    const enemy = this.findNearestBotEnemy(bot);
-    if (!enemy) return;
-    if (!isOnlineActorTargetableByBot(bot.team, enemy, this.botGoalAxis, this.botGoalSigns)) return;
-    const dx = enemy.posX - bot.posX;
-    const dy = enemy.posY - bot.posY;
-    const dz = enemy.posZ - bot.posZ;
-    const distSq = dx * dx + dy * dy + dz * dz;
-    if (distSq > p.maxRange * p.maxRange) return;
-    const dist = Math.sqrt(distSq);
-    const nx = dx / dist + (Math.random() - 0.5) * p.angleNoise;
-    const ny = dy / dist + (Math.random() - 0.5) * p.angleNoise;
-    const nz = dz / dist + (Math.random() - 0.5) * p.angleNoise;
-    const nLen = Math.hypot(nx, ny, nz) || 1;
-    const shotEvent: ShotEventMessage = {
-      ownerId: bot.id,
-      team: bot.team,
-      originX: bot.posX,
-      originY: bot.posY,
-      originZ: bot.posZ,
-      dirX: nx / nLen,
-      dirY: ny / nLen,
-      dirZ: nz / nLen,
-    };
-    this.broadcast("shot_event", shotEvent);
-    const rangeFactor = 1 - (dist / p.maxRange) * 0.5;
-    const hitChance = (0.35 + p.tier * 0.1) * rangeFactor;
-    if (Math.random() > hitChance || enemy.frozen) return;
-    if (enemy.isBot) {
-      enemy.frozen = true;
-      enemy.phase = "FROZEN";
-      enemy.deaths += 1;
-    } else {
-      const frozen = applyHitToOnlineActor(enemy, "body");
-      if (!frozen) return;
-    }
-    enemy.frozenTimer = BOT_RESPAWN_SECONDS;
-    bot.kills = Math.min(MAX_KILLS, bot.kills + 1);
-    const freezeEvent: FreezeEventMessage = {
-      targetId: enemy.id,
-      killerName: bot.name,
-      killerTeam: bot.team,
-      victimName: enemy.name,
-      victimTeam: enemy.team,
-    };
-    this.broadcast("freeze_event", freezeEvent);
-    this.checkFullFreezeWin();
   }
 
   // ── Lobby helpers ───────────────────────────────────────────────────────────
@@ -1132,12 +1013,6 @@ function normalizeDirection(
   };
 }
 
-function breachRoomCenter(goalAxis: "x" | "y" | "z", sign: 1 | -1): { x: number; y: number; z: number } {
-  const center = { x: 0, y: 0, z: 0 };
-  center[goalAxis] = sign * (ARENA_SIZE / 2 + BREACH_ROOM_D / 2);
-  return center;
-}
-
 function breachExitYaw(axis: "x" | "y" | "z", openSign: 1 | -1): number {
   const dx = axis === "x" ? openSign : 0;
   const dz = axis === "z" ? openSign : 0;
@@ -1159,139 +1034,3 @@ function makeServerArenaQuery(
   };
 }
 
-interface BotPersonality {
-  tier: number;
-  launchSpeed: number;
-  fireDelay: number;
-  angleNoise: number;
-  maxRange: number;
-}
-
-function botPersonality(idHash: number): BotPersonality {
-  const tier = idHash % 5;
-  return {
-    tier,
-    launchSpeed: 6 + tier * 2,      // 6..14
-    fireDelay: 3.0 - tier * 0.4,    // 3.0..1.4s
-    angleNoise: 0.45 - tier * 0.08, // 0.45..0.13
-    maxRange: 15 + tier * 5,        // 15..35
-  };
-}
-
-function botIdHash(id: string): number {
-  let h = 0;
-  for (let i = 0; i < id.length; i++) {
-    h = (h + id.charCodeAt(i)) | 0;
-  }
-  return Math.abs(h);
-}
-
-function botIntegrateZeroG(actor: ActorState, dt: number): void {
-  integrateZeroGActor(actor, dt);
-}
-
-function botBounceArena(actor: ActorState, goalAxis: "x" | "z"): void {
-  bounceActorInArena(actor, goalAxis);
-  return;
-  const half = ARENA_SIZE / 2 - PLAYER_RADIUS;
-  const perpAxis: "x" | "z" = goalAxis === "x" ? "z" : "x";
-
-  // Y axis — always solid
-  if (actor.posY < -half) { actor.posY = -half; actor.velY = Math.abs(actor.velY); }
-  else if (actor.posY > half) { actor.posY = half; actor.velY = -Math.abs(actor.velY); }
-
-  // Perp axis — always solid
-  if (perpAxis === "x") {
-    if (actor.posX < -half) { actor.posX = -half; actor.velX = Math.abs(actor.velX); }
-    else if (actor.posX > half) { actor.posX = half; actor.velX = -Math.abs(actor.velX); }
-  } else {
-    if (actor.posZ < -half) { actor.posZ = -half; actor.velZ = Math.abs(actor.velZ); }
-    else if (actor.posZ > half) { actor.posZ = half; actor.velZ = -Math.abs(actor.velZ); }
-  }
-
-  // Goal axis — portal openings on both walls; breach room back wall at ±(ARENA_SIZE/2 + BREACH_ROOM_D)
-  const perpPos = perpAxis === "x" ? actor.posX : actor.posZ;
-  const inPortal = Math.abs(actor.posY) < BREACH_ROOM_H / 2 - PLAYER_RADIUS
-    && Math.abs(perpPos) < BREACH_ROOM_W / 2 - PLAYER_RADIUS;
-  const maxDepth = ARENA_SIZE / 2 + BREACH_ROOM_D - PLAYER_RADIUS;
-
-  if (goalAxis === "x") {
-    if (actor.posX < -half) {
-      if (inPortal) {
-        if (actor.posX < -maxDepth) { actor.posX = -maxDepth; actor.velX = Math.abs(actor.velX); }
-      } else {
-        actor.posX = -half; actor.velX = Math.abs(actor.velX);
-      }
-    } else if (actor.posX > half) {
-      if (inPortal) {
-        if (actor.posX > maxDepth) { actor.posX = maxDepth; actor.velX = -Math.abs(actor.velX); }
-      } else {
-        actor.posX = half; actor.velX = -Math.abs(actor.velX);
-      }
-    }
-  } else {
-    if (actor.posZ < -half) {
-      if (inPortal) {
-        if (actor.posZ < -maxDepth) { actor.posZ = -maxDepth; actor.velZ = Math.abs(actor.velZ); }
-      } else {
-        actor.posZ = -half; actor.velZ = Math.abs(actor.velZ);
-      }
-    } else if (actor.posZ > half) {
-      if (inPortal) {
-        if (actor.posZ > maxDepth) { actor.posZ = maxDepth; actor.velZ = -Math.abs(actor.velZ); }
-      } else {
-        actor.posZ = half; actor.velZ = -Math.abs(actor.velZ);
-      }
-    }
-  }
-}
-
-interface OnlineBotTargetCandidate {
-  team: 0 | 1;
-  frozen: boolean;
-  phase: string;
-  posX: number;
-  posY: number;
-  posZ: number;
-}
-
-export function isOnlineActorTargetableByBot(
-  botTeam: 0 | 1,
-  actor: OnlineBotTargetCandidate,
-  goalAxis: "x" | "z",
-  goalSigns: { team0: 1 | -1; team1: 1 | -1 },
-): boolean {
-  if (actor.team === botTeam || actor.frozen || actor.phase === "RESPAWNING") {
-    return false;
-  }
-
-  return getOnlineActorBreachTeam(actor, goalAxis, goalSigns) === null;
-}
-
-export function getOnlineActorBreachTeam(
-  actor: Pick<OnlineBotTargetCandidate, "posX" | "posY" | "posZ">,
-  goalAxis: "x" | "z",
-  goalSigns: { team0: 1 | -1; team1: 1 | -1 },
-): 0 | 1 | null {
-  return isPointInsideBreachRoom(actor, breachRoomCenter(goalAxis, goalSigns.team0), goalAxis)
-    ? 0
-    : isPointInsideBreachRoom(actor, breachRoomCenter(goalAxis, goalSigns.team1), goalAxis)
-      ? 1
-      : null;
-}
-
-function isPointInsideBreachRoom(
-  pos: Pick<OnlineBotTargetCandidate, "posX" | "posY" | "posZ">,
-  center: { x: number; y: number; z: number },
-  goalAxis: "x" | "z",
-): boolean {
-  if (Math.abs(pos.posY - center.y) >= BREACH_ROOM_H / 2) return false;
-
-  const depthPos = goalAxis === "x" ? pos.posX : pos.posZ;
-  const depthCenter = goalAxis === "x" ? center.x : center.z;
-  if (Math.abs(depthPos - depthCenter) >= BREACH_ROOM_D / 2) return false;
-
-  const perpPos = goalAxis === "x" ? pos.posZ : pos.posX;
-  const perpCenter = goalAxis === "x" ? center.z : center.x;
-  return Math.abs(perpPos - perpCenter) < BREACH_ROOM_W / 2;
-}
